+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+## v6.26.1 (9 January 2024)
+
+### Fixes
+
+* Fix deprecation warning from Sidekiq error handler
+  | [#796](https://github.com/bugsnag/bugsnag-ruby/pull/796)
+  | [fukayatsu](https://github.com/fukayatsu)
+* Fix Resque integration when failure backend is already `Resque::Failure::Multiple`
+  | [#803](https://github.com/bugsnag/bugsnag-ruby/pull/803)
+  | [sj26](https://github.com/sj26)
+* Redact URLs in automatic Rails breadcrumbs
+  | [#806](https://github.com/bugsnag/bugsnag-ruby/pull/806)
+
 ## v6.26.0 (19 July 2023)
 
 ### Enhancements